Refactor trip-related components and remove deprecated place functionality
- Updated `.vscode` configuration files for improved project setup.
- Adjusted `cwd` and `program` paths in `launch.json` for the new feature structure.
- Refactored `signup` component to use nullish coalescing for safer property access.
- Corrected import paths in `ManualTripCreate`, `SectionHeader`, and `TripItemCard` components to reflect the new directory structure.
- Removed unused place-related files and functionality from the `trips` feature, enhancing codebase maintainability.
