+++ .fttemplates/CleanArchitectureFeature/.ftsettings.json
@@ -1,0 +1,4 @@
+{
+  "name": "Clean Architecture Feature",
+  "customVariables": []
+}
