+++ components/CreateTripComponents/ManualTripCreate.tsx
@@ -3,7 +3,7 @@ import { useManualTripStore } from '@/store/manualTripStore'
 import React, { useEffect, useState } from 'react'
 import { StyleSheet } from 'react-native'
 import { Button, View } from 'react-native-ui-lib'
-import TripPlanner from '../../features/trips/presentation/components/TripPlanner'
+import TripPlanner from '../../features/trip/presentation/components/TripPlanner'
 import HorizontalDatePicker from './HorizontalDatePicker'
 
 type ManualTripCreateProps = {
