+++ features/place/data/PlaceRepoImpl.ts
@@ -1,9 +1,9 @@
-import { Place } from '@/features/trips/domain/models/Place'
+import { Place } from '@/features/place/domain/models/Place'
+import { PlaceApi } from '../infrastructure/api/PlaceApi'
 import {
   GetPlacesParams,
   PlaceRepository,
 } from '../domain/repositories/IPlaceRepository'
-import { PlaceApi } from '../infrastructure/api/PlaceApi'
 
 export class PlaceRepoImpl implements PlaceRepository {
   async getPlaces(params: GetPlacesParams): Promise<Place[]> {
