+++ features/place/domain/models/Place.ts
@@ -1,0 +1,11 @@
+export type Place = {
+  en_name: string
+  en_properties: string[]
+  en_type: string
+  id: string
+  lat: number
+  long: number
+  vi_name: string
+  vi_properties: string[]
+  vi_type: string
+}
