+++ features/place/presentation/state/usePlaces.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query'
+import { PlaceRepoImpl } from '../../data/PlaceRepoImpl'
 import { GetPlacesParams } from '../../domain/repositories/IPlaceRepository'
 import { GetAllPlaces } from '../../domain/usecases/GetAllPlaces'
-import { PlaceRepoImpl } from '../../data/PlaceRepoImpl'
 
 export const usePlaces = (params: GetPlacesParams) => {
   const getAllPlaces = new GetAllPlaces(new PlaceRepoImpl())
